refactored common fixtures into conftest.py file
